Harden the terminal client, resumes and quoting
- A failed or dying control-mode attach was invisible: stderr
  went to /dev/null and the `%exit` reason was dropped, leaving a
  permanently blank pane; exits now report to the Errors tab,
  print into the pane and fire the termination callback, with
  clean detaches staying silent
- The attach also heals `history-limit` on servers born before
  the config carried it, so seeded scrollback is deep
- Option-drag block selection now hit-tests the window, so a
  mounted but hidden pane cannot steal drags or the clipboard
- Resuming a worktree without a recorded resume id launches a
  fresh session instead of re-feeding the original prompt, which
  re-ran the whole task against the modified worktree
- Shell quoting and the scrubbed process environment collapse to
  one implementation each (`shellQuoted`, `ProcessEnvironment`),
  with the quoting under test as the paste-delivery successor
