+++ Sources/AgentIDEData/AgentRunner.swift
@@ -55,9 +55,7 @@ extension AgentRunner {
         let trimmed = extraArguments.trimmingCharacters(in: .whitespaces)
         var joined = trimmed.isEmpty ? base : base + " " + trimmed
         if let promptFile {
-            // The path is single-quoted for the shell, so any single
-            // quotes in it must be escaped out of the quoting.
-            joined += " \"$(cat '" + promptFile.replacing("'", with: "'\\''") + "')\""
+            joined += " \"$(cat " + promptFile.shellQuoted + ")\""
         }
         return joined
     }
